fixing groups and services
